requirements.txt
langchain
langchain-groq
langchain-community
langchain-text-splitters
faiss-cpu
sentence-transformers
python-dotenv
